requirements.txt
astroid==3.3.5
dill==0.3.9
isort==5.13.2
mccabe==0.7.0
pip-autoremove==0.10.0
platformdirs==4.3.6
pydot==3.0.2
pylint==3.3.1
pyparsing==3.2.0
pytomutil-elunico==0.0.5
setuptools==75.3.0
spllib-elunico==0.0.4
tomlkit==0.13.2
